+++ system/platform-core/src/main/java/org/platformlambda/core/annotations/PreLoad.java
@@ -24,7 +24,11 @@ import java.lang.annotation.*;
  * This indicates the class is a service to be preloaded.
  * (for a class to be preloaded, it must use a default constructor without arguments)
  * <p>
- * envInstances overrides instances from multiple sources.
+ * customSerializer is optional. It allows you to apply a pre-configured ObjectMapper suitable for
+ * your user function. It should point to a fully qualified classpath to a class implementing
+ * the CustomSerializer interface.
+ * <p>
+ * envInstances is optional. It overrides the "instances" parameter using environment or application.properties.
  * <p>
  * 1. To get value from an environment variable, use this format ${ENV_VAR_NAME:defaultValue}.
  * 2. To get value from application.properties or application.yml, just set it to the parameter name.
@@ -36,8 +40,10 @@ import java.lang.annotation.*;
 public @interface PreLoad {
 
     String route();
+    String customSerializer() default "";
     int instances() default 1;
     String envInstances() default "";
+
     boolean isPrivate() default true;
 
 }