+++ CHANGELOG.md
@@ -16,6 +16,7 @@ For release notes before version 3.1, please refer to https://github.com/Accentu
 
 1. AutoStart to run application as Spring Boot if the rest-spring-3 library is packaged in app
 2. Configurable "Event over HTTP" - automatic forward events over HTTP using a configuration
+3. Support user defined serializer with PreLoad annotation and platform API
 
 ### Removed